Health Parameter on Create Character
Noticed health was assigned by a parameter and then checked if it was
correct. I thought it would be better if health was set when the
character type was selected. However I didn't want to change things that
may require a refactor in many other places so I left a comment.